requirements.txt
Django==3.2.16
Pillow==9.3.0

django-allauth==0.51.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0

flake8==6.0.0
isort==5.10.1

stripe==5.0.0

redis==4.3.5
django-redis==5.2.0

celery==5.2.7
psycopg2==2.9.5

djangorestframework~=3.14.0